NMS-9288: Topology on Map is not properly displayed
More refactoring
Move all searchProvider method to the ACL provider class
Move all private classes to their own class file
Cleaned more other methods
