+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -62,10 +62,7 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         <property name="nodeDao" ref="nodeDao" />
         <property name="snmpInterfaceDao" ref="snmpInterfaceDao" />
         <property name="ipInterfaceDao" ref="ipInterfaceDao" />
-        <property name="topologyDao" ref="topologyDao"/>
         <property name="filterManager" ref="filterManager" />
-        <property name="addNodeWithoutLink" value="true"/>
-        <property name="linkdHopCriteriaFactory" ref="linkdHopCriteriaFactory" />
         <property name="lldpElementDao" ref="lldpElementDao" />
         <property name="lldpLinkDao" ref="lldpLinkDao" />
         <property name="cdpLinkDao" ref="cdpLinkDao"/>
@@ -74,8 +71,9 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         <property name="isisLinkDao" ref="isisLinkDao" />
         <property name="isisElementDao" ref="isisElementDao" />
         <property name="bridgeTopologyDao" ref="bridgeTopologyDao" />
-        <property name="topologyProviderInfo" ref="enlinkdTopologyProviderInfo"/>
         <property name="ipNetToMediaDao" ref="ipNetToMediaDao"/>
+        <property name="topologyProviderInfo" ref="enlinkdTopologyProviderInfo"/>
+        <property name="topologyDao" ref="topologyDao"/>
     </bean>
 
     <!--
@@ -104,7 +102,7 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         </service-properties>
     </service>
 
-    <service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="topologyProvider" />
+    <service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="topologyACLProvider" />
 
     <!-- EnhancedLinkd StatusProviders Service and CheckedOperation -->
     <bean id="linkdEdgeStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdEdgeStatusProvider">
